+++ framework/Code/DKDistortionTransform.mm
@@ -23,14 +23,14 @@ extern "C" {
 #endif
 
 #pragma mark Static Functions
-static inline float	MMul( float a, float b, float c, float d )
+static inline CGFloat	MMul( CGFloat a, CGFloat b, CGFloat c, CGFloat d )
 {
 	return a*d-b*c;
 }
 
-static inline void	VP(float *px, float *py, float x1, float y1, float x2, float y2, float x3, float y3, float x4, float y4)
+static inline void	VP(CGFloat *px, CGFloat *py, CGFloat x1, CGFloat y1, CGFloat x2, CGFloat y2, CGFloat x3, CGFloat y3, CGFloat x4, CGFloat y4)
 {
-  float d = MMul(x1-x2,y1-y2,x3-x4,y3-y4);
+  CGFloat d = MMul(x1-x2,y1-y2,x3-x4,y3-y4);
   
   if (d==0.0f)
     d = 1.0f;
@@ -46,11 +46,13 @@ static NSPoint	Map( NSPoint inPoint, NSSize sourceSize, NSPoint quad[4])
 	
 	NSPoint		p;
 	
-	VP( (float*) &p.x, (float*) &p.y,
-	   (float) ((sourceSize.height-inPoint.y)*quad[0].x + (inPoint.y)*quad[3].x)/sourceSize.height, (float) ((sourceSize.height-inPoint.y)*quad[0].y + inPoint.y*quad[3].y)/sourceSize.height,
-	   (float) ((sourceSize.height-inPoint.y)*quad[1].x + (inPoint.y)*quad[2].x)/sourceSize.height, (float) ((sourceSize.height-inPoint.y)*quad[1].y + inPoint.y*quad[2].y)/sourceSize.height,
-	   (float) ((sourceSize.width-inPoint.x)*quad[0].x + (inPoint.x)*quad[1].x)/sourceSize.width, (float) ((sourceSize.width-inPoint.x)*quad[0].y + inPoint.x*quad[1].y)/sourceSize.width,
-	   (float) ((sourceSize.width-inPoint.x)*quad[3].x + (inPoint.x)*quad[2].x)/sourceSize.width, (float) ((sourceSize.width-inPoint.x)*quad[3].y + inPoint.x*quad[2].y)/sourceSize.width);
+#warning 64BIT: Inspect pointer casting
+#warning 64BIT: Inspect pointer casting
+	VP( (CGFloat *)&p.x, (CGFloat *)&p.y,
+	   (CGFloat) ((sourceSize.height-inPoint.y)*quad[0].x + (inPoint.y)*quad[3].x)/sourceSize.height, (CGFloat) ((sourceSize.height-inPoint.y)*quad[0].y + inPoint.y*quad[3].y)/sourceSize.height,
+	   (CGFloat) ((sourceSize.height-inPoint.y)*quad[1].x + (inPoint.y)*quad[2].x)/sourceSize.height, (CGFloat) ((sourceSize.height-inPoint.y)*quad[1].y + inPoint.y*quad[2].y)/sourceSize.height,
+	   (CGFloat) ((sourceSize.width-inPoint.x)*quad[0].x + (inPoint.x)*quad[1].x)/sourceSize.width, (CGFloat) ((sourceSize.width-inPoint.x)*quad[0].y + inPoint.x*quad[1].y)/sourceSize.width,
+	   (CGFloat) ((sourceSize.width-inPoint.x)*quad[3].x + (inPoint.x)*quad[2].x)/sourceSize.width, (CGFloat) ((sourceSize.width-inPoint.x)*quad[3].y + inPoint.x*quad[2].y)/sourceSize.width);
 	
 	return p;
 }
@@ -126,7 +128,7 @@ static NSPoint	Map( NSPoint inPoint, NSSize sourceSize, NSPoint quad[4])
 	// returns a rect bounding the envelope points
 	
 	NSRect	r = NSZeroRect;
-	int		i;
+	NSInteger		i;
 	
 	for( i = 0; i < 4; ++i )
 		r = NSUnionRect( r, NSInsetRect( NSRectFromTwoPoints( m_q[i], m_q[i] ), -1.0, -1.0 ));
@@ -136,7 +138,7 @@ static NSPoint	Map( NSPoint inPoint, NSSize sourceSize, NSPoint quad[4])
 
 
 #pragma mark -
-- (void)			offsetByX:(float) dx byY:(float) dy
+- (void)			offsetByX:(CGFloat) dx byY:(CGFloat) dy
 {
 	m_q[0].x += dx;
 	m_q[1].x += dx;
@@ -149,7 +151,7 @@ static NSPoint	Map( NSPoint inPoint, NSSize sourceSize, NSPoint quad[4])
 }
 
 
-- (void)			shearHorizontallyBy:(float) dx
+- (void)			shearHorizontallyBy:(CGFloat) dx
 {
 	m_q[0].x += dx;
 	m_q[1].x += dx;
@@ -158,7 +160,7 @@ static NSPoint	Map( NSPoint inPoint, NSSize sourceSize, NSPoint quad[4])
 }
 
 
-- (void)			shearVerticallyBy:(float) dy
+- (void)			shearVerticallyBy:(CGFloat) dy
 {
 	m_q[0].y -= dy;
 	m_q[3].y -= dy;
@@ -167,7 +169,7 @@ static NSPoint	Map( NSPoint inPoint, NSSize sourceSize, NSPoint quad[4])
 }
 
 
-- (void)			differentialPerspectiveBy:(float) delta
+- (void)			differentialPerspectiveBy:(CGFloat) delta
 {
 	m_q[0].y += delta;
 	m_q[1].y -= delta;
@@ -246,7 +248,7 @@ static NSPoint	Map( NSPoint inPoint, NSSize sourceSize, NSPoint quad[4])
 	NSBezierPath*		newPath = [path copy];
 	NSPoint				ap[3];
 	NSBezierPathElement	elem;
-	int					i, ec;
+	NSInteger					i, ec;
 	NSRect				bounds = [path controlPointBounds];
 	
 	[newPath removeAllPoints];
