+++ framework/Code/DKGridLayer.m
@@ -16,6 +16,7 @@
 #import "DKDrawingView.h"
 #import "NSColor+DKAdditions.h"
 #import "LogEvent.h"
+#include <tgmath.h>
 
 
 #pragma mark Contants (Non-localized)
@@ -722,7 +723,7 @@ static NSColor*		sMajorColour = nil;
 	CGFloat		newLeft, newTop, newRight, newBottom;
 	
 	CGFloat dim = paper.width - marg;
-	CGFloat rem = fmodf( dim, q );
+	CGFloat rem = fmod( dim, q );
 	
 	if( rem < 0.001 )
 		rem = 0;
@@ -732,7 +733,7 @@ static NSColor*		sMajorColour = nil;
 	
 	marg = [[self drawing] topMargin] + [[self drawing] bottomMargin];
 	dim = paper.height - marg;
-	rem = fmodf( dim, q );
+	rem = fmod( dim, q );
 
 	if( rem < 0.001 )
 		rem = 0;
@@ -916,14 +917,14 @@ static NSColor*		sMajorColour = nil;
 - (NSPoint)					nearestGridIntersectionToPoint:(NSPoint) p
 {
 	CGFloat		dd = [self divisionDistance];
-	CGFloat		rem = fmodf( p.x - [[self drawing] leftMargin], dd );
+	CGFloat		rem = fmod( p.x - [[self drawing] leftMargin], dd );
 	
 	if ( rem > dd * 0.5f )
 		p.x += ( dd - rem );
 	else
 		p.x -= rem;
 	
-	rem = fmodf( p.y - [[self drawing] topMargin], dd );
+	rem = fmod( p.y - [[self drawing] topMargin], dd );
 	
 	if ( rem > dd * 0.5f )
 		p.y += ( dd - rem );
@@ -960,7 +961,7 @@ static NSColor*		sMajorColour = nil;
 	else
 	{
 		divs = [self divisionDistance];
-		rem = fmodf( size.width, divs );
+		rem = fmod( size.width, divs );
 		if ( rem > divs / 2.0 )
 			size.width += ( divs - rem );
 		else
@@ -971,7 +972,7 @@ static NSColor*		sMajorColour = nil;
 		size.height = interior.size.height;
 	else
 	{
-		rem = fmodf( size.height, divs );
+		rem = fmod( size.height, divs );
 		if ( rem > divs / 2.0 )
 			size.height += ( divs - rem );
 		else
@@ -1008,7 +1009,7 @@ static NSColor*		sMajorColour = nil;
 	else
 	{
 		divs = [self spanDistance];
-		rem = fmodf( size.width, divs );
+		rem = fmod( size.width, divs );
 		if ( rem > divs / 2.0 )
 			size.width += ( divs - rem );
 		else
@@ -1020,7 +1021,7 @@ static NSColor*		sMajorColour = nil;
 	else
 	{
 		divs = [self spanDistance];
-		rem = fmodf( size.height, divs );
+		rem = fmod( size.height, divs );
 		if ( rem > divs / 2.0 )
 			size.height += ( divs - rem );
 		else
@@ -1711,6 +1712,7 @@ static NSColor*		sMajorColour = nil;
 	if ( action == @selector( copy: ))
 	{
 		NSString* title = NSLocalizedString(@"Copy %@", nil);
+#warning 64BIT: Check formatting arguments
 		[item setTitle:[NSString stringWithFormat:title, [self layerName]]];
 		return YES;
 	}
